+++ R/helpymchelperson.R
@@ -9,3 +9,15 @@
   return(x)
 
 }
+
+#' Extract months from date objects
+#' @param date a date
+#' @keywords internal
+#' @export
+#' @noRd
+
+.pshf_month <- function(date) {
+  x <- as.POSIXlt(date)$mon + 1 # Importantly, Jan = 0 and Dec = 11
+  return(x)
+
+}